Geeks and leetcode

+++ src/main/java/SaveGowtham.java
@@ -1,5 +1,7 @@
-public class SaveGowtham{
-public static void main(String[] args) {
+import java.util.Scanner;
+
+public class SaveGowtham {
+    public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
         int t = sc.nextInt();
         while (t > 0) {
@@ -8,7 +10,7 @@ public static void main(String[] args) {
             for (int i = 0; i <n; i++) {
                 a[i]=sc.nextInt();
             }
-            long maximum=0L;
+            int maximum=0;
             for (int i = 0; i <n; i++) {
                 for (int k = i + 1; k < n; k++) {
                     if (a[k] > a[i]) {
